VB-650 Refactor Integration Test and Unit tests
