+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/VisitSessionsControllerTest.kt
@@ -1,27 +1,30 @@
-package uk.gov.justice.digital.hmpps.visitscheduler.controller
+package uk.gov.justice.digital.hmpps.visitscheduler.integration
 
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Import
-import uk.gov.justice.digital.hmpps.visitscheduler.config.TestClockConfiguration
-import uk.gov.justice.digital.hmpps.visitscheduler.helper.sessionTemplate
-import uk.gov.justice.digital.hmpps.visitscheduler.helper.sessionTemplateCreator
-import uk.gov.justice.digital.hmpps.visitscheduler.helper.sessionTemplateDeleter
-import uk.gov.justice.digital.hmpps.visitscheduler.helper.visitCreator
-import uk.gov.justice.digital.hmpps.visitscheduler.helper.visitDeleter
-import uk.gov.justice.digital.hmpps.visitscheduler.integration.IntegrationTestBase
-import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency
+import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency.DAILY
+import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency.MONTHLY
+import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency.SINGLE
+import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency.WEEKLY
 import uk.gov.justice.digital.hmpps.visitscheduler.model.VisitRestriction
 import uk.gov.justice.digital.hmpps.visitscheduler.model.VisitStatus
 import uk.gov.justice.digital.hmpps.visitscheduler.model.VisitType
 import uk.gov.justice.digital.hmpps.visitscheduler.repository.SessionTemplateRepository
 import uk.gov.justice.digital.hmpps.visitscheduler.repository.VisitRepository
+import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.TestClockConfiguration
+import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.sessionTemplate
+import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.sessionTemplateCreator
+import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.sessionTemplateDeleter
+import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.visitCreator
+import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.visitDeleter
+import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.integration.IntegrationTestBase
 import java.time.LocalDate
 
 @Import(TestClockConfiguration::class)
-class VisitSessionsControllerIntTest : IntegrationTestBase() {
+class VisitSessionsControllerTest : IntegrationTestBase() {
 
   @Autowired
   private lateinit var sessionTemplateRepository: SessionTemplateRepository
@@ -41,7 +44,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
       repository = sessionTemplateRepository,
       sessionTemplate = sessionTemplate(
         startDate = LocalDate.parse("2021-01-08"),
-        frequency = SessionFrequency.SINGLE,
+        frequency = SINGLE,
         restrictions = "Only B wing"
       )
     ).save()
@@ -71,7 +74,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
       sessionTemplate = sessionTemplate(
         startDate = LocalDate.parse("2021-01-08"),
         expiryDate = LocalDate.parse("2021-01-14"),
-        frequency = SessionFrequency.DAILY,
+        frequency = DAILY,
       )
     ).save()
 
@@ -89,7 +92,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
       repository = sessionTemplateRepository,
       sessionTemplate = sessionTemplate(
         startDate = LocalDate.parse("2021-01-08"),
-        frequency = SessionFrequency.WEEKLY,
+        frequency = WEEKLY,
         restrictions = "Only B wing"
       )
     ).save()
@@ -117,7 +120,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
       repository = sessionTemplateRepository,
       sessionTemplate = sessionTemplate(
         startDate = LocalDate.parse("2021-01-08"),
-        frequency = SessionFrequency.MONTHLY
+        frequency = MONTHLY
       )
     ).save()
 
@@ -138,7 +141,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
       sessionTemplate = sessionTemplate(
         startDate = LocalDate.parse("2020-01-01"),
         expiryDate = LocalDate.parse("2020-06-01"),
-        frequency = SessionFrequency.WEEKLY,
+        frequency = WEEKLY,
         restrictions = "Only B wing"
       )
     ).save()
@@ -158,7 +161,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
       sessionTemplate = sessionTemplate(
         startDate = LocalDate.parse("2022-01-01"),
         expiryDate = LocalDate.parse("2022-06-01"),
-        frequency = SessionFrequency.WEEKLY,
+        frequency = WEEKLY,
         restrictions = "Only B wing"
       )
     ).save()
@@ -180,7 +183,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
         startDate = dateTime.toLocalDate(),
         startTime = dateTime.toLocalTime(),
         endTime = dateTime.plusHours(1).toLocalTime(),
-        frequency = SessionFrequency.SINGLE
+        frequency = SINGLE
       )
     ).save()
     visitCreator(visitRepository)
@@ -218,7 +221,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
         startDate = dateTime.toLocalDate(),
         startTime = dateTime.toLocalTime(),
         endTime = dateTime.plusHours(1).toLocalTime(),
-        frequency = SessionFrequency.SINGLE
+        frequency = SINGLE
       )
     ).save()
     visitCreator(visitRepository)
@@ -259,7 +262,7 @@ class VisitSessionsControllerIntTest : IntegrationTestBase() {
         startDate = dateTime.toLocalDate(),
         startTime = dateTime.toLocalTime(),
         endTime = dateTime.plusHours(1).toLocalTime(),
-        frequency = SessionFrequency.SINGLE
+        frequency = SINGLE
       )
     ).save()
     visitCreator(visitRepository)
