+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/test_setup/helper/EntityTestBuilder.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.visitscheduler.helper
+package uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper
 
 import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency
 import uk.gov.justice.digital.hmpps.visitscheduler.model.VisitType
